requirements.txt
contourpy==1.1.1
control==0.9.4
cycler==0.12.1
fonttools==4.43.1
h5py==3.10.0
kiwisolver==1.4.5
matlabengine==23.2.1
matplotlib==3.8.0
mypy==1.6.1
mypy-extensions==1.0.0
numpy==1.26.1
packaging==23.2
pandas==2.1.1
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
scipy==1.11.3
six==1.16.0
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
